Comportamientos nuevos de anomalias con acciones
Script para objetos interactivos anomalías y para enemigos
