+++ CameraPrototype/Assets/_Scripts/Anomalies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: baaa55ba2783d0e4a9c8013e10486f86
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
