+++ CameraPrototype/Assets/_Scripts/Anomalies/AnomalyInteractiveObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7faed95056743b42929f5dd1c41d5e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
