--- CameraPrototype/Assets/_Scripts/Camera/Anomalies Camera/AnomaliesData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7d61abdabda17524bbba73d4387d096e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
