--- CameraPrototype/Assets/_Scripts/Camera/Anomalies Camera/AnomaliesLink.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b64a97ad9099f4d42b6c2d1a8e68b18d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
